3-Orchestration/requirements.txt
aiofiles==0.8.0
aiohttp==3.8.1
aiohttp-cors==0.7.0
aiosignal==1.2.0
aiosqlite==0.17.0
alembic==1.8.0
anyio==3.6.1
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgi-lifespan==1.0.1
asgiref==3.5.2
asttokens==2.0.5
async-timeout==4.0.2
asyncpg==0.25.0
attrs==21.4.0
azure-core==1.24.1
azure-storage-blob==12.12.0
backcall==0.2.0
beautifulsoup4==4.11.1
bleach==5.0.0
blessed==1.19.1
boto3==1.24.2
botocore==1.27.2
cachetools==5.2.0
certifi==2022.5.18.1
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.3
cloudpickle==2.1.0
colorful==0.5.4
commonmark==0.9.1
coolname==1.1.0
cramjam==2.5.0
croniter==1.3.5
cryptography==37.0.2
cycler==0.11.0
dask==2022.5.2
databricks-cli==0.16.6
debugpy==1.6.0
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.4
distributed==2022.5.2
docker==5.0.3
entrypoints==0.4
executing==0.8.3
fastapi==0.78.0
fastjsonschema==2.15.3
fastparquet==0.8.1
filelock==3.7.1
Flask==2.1.2
fonttools==4.33.3
frozenlist==1.3.0
fsspec==2022.5.0
future==0.18.2
gitdb==4.0.9
GitPython==3.1.27
google-api-core==2.8.1
google-auth==2.6.6
google-cloud-core==2.3.0
google-cloud-storage==2.3.0
google-crc32c==1.3.0
google-resumable-media==2.3.3
googleapis-common-protos==1.56.2
gpustat==1.0.0b1
greenlet==1.1.2
grpcio==1.43.0
gunicorn==20.1.0
h11==0.12.0
HeapDict==1.0.1
httpcore==0.15.0
httpx==0.23.0
hyperopt==0.2.7
idna==3.3
importlib-metadata==4.11.4
ipykernel==6.13.0
ipython==8.4.0
ipython-genutils==0.2.0
ipywidgets==7.7.0
isodate==0.6.1
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.1.2
jmespath==1.0.0
joblib==1.1.0
jsonschema==4.6.0
jupyter==1.0.0
jupyter-client==7.3.2
jupyter-console==6.4.3
jupyter-core==4.10.0
jupyterlab-pygments==0.2.2
jupyterlab-widgets==1.1.0
kiwisolver==1.4.2
kubernetes==23.6.0
locket==1.0.0
Mako==1.2.0
MarkupSafe==2.1.1
matplotlib==3.5.2
matplotlib-inline==0.1.3
mistune==0.8.4
mlflow==1.26.1
msgpack==1.0.4
msrest==0.6.21
multidict==6.0.2
nbclient==0.6.4
nbconvert==6.5.0
nbformat==5.4.0
nest-asyncio==1.5.5
networkx==2.8.3
notebook==6.4.11
numpy==1.22.4
nvidia-ml-py3==7.352.0
oauthlib==3.2.0
opencensus==0.9.0
opencensus-context==0.1.2
packaging==21.3
pandas==1.4.2
pandocfilters==1.5.0
parso==0.8.3
partd==1.2.0
pendulum==2.1.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.1.1
platformdirs==2.5.2
prefect==2.0b5
prometheus-client==0.13.1
prometheus-flask-exporter==0.20.2
prompt-toolkit==3.0.29
protobuf==3.20.1
psutil==5.9.1
ptyprocess==0.7.0
pure-eval==0.2.2
py-spy==0.3.12
py4j==0.10.9.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydantic==1.9.1
Pygments==2.12.0
PyJWT==2.4.0
pyparsing==3.0.9
pyrsistent==0.18.1
python-dateutil==2.8.2
python-slugify==6.1.2
pytz==2022.1
pytzdata==2020.1
PyYAML==6.0
pyzmq==23.1.0
qtconsole==5.3.1
QtPy==2.1.0
querystring-parser==1.2.4
ray==1.12.1
readchar==3.0.5
requests==2.27.1
requests-oauthlib==1.3.1
rfc3986==1.5.0
rich==12.4.4
rsa==4.8
s3transfer==0.6.0
scikit-learn==1.1.1
scipy==1.8.1
seaborn==0.11.2
Send2Trash==1.8.0
six==1.16.0
smart-open==6.0.0
smmap==5.0.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
SQLAlchemy==1.4.37
sqlparse==0.4.2
stack-data==0.2.0
starlette==0.19.1
tabulate==0.8.9
tblib==1.7.0
terminado==0.15.0
text-unidecode==1.3
threadpoolctl==3.1.0
tinycss2==1.1.1
toml==0.10.2
toolz==0.11.2
tornado==6.1
tqdm==4.64.0
traitlets==5.2.2.post1
typer==0.4.1
typing_extensions==4.2.0
urllib3==1.26.9
uvicorn==0.17.6
virtualenv==20.14.1
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.3.2
Werkzeug==2.1.2
wget==3.2
widgetsnbextension==3.6.0
xgboost==1.6.1
yarl==1.7.2
zict==2.2.0
zipp==3.8.0
